+++ CHANGELOG.md
@@ -2,5 +2,14 @@
 
 ## [Unreleased]
 
+
+## [1.0.0] - 2021-04-20
+
+- linux: changed default openType to chromium-browser (because of Firefox crash)
+- added setting openType (lets you specify specific browser)
+- added log
+- added browser not installed message
+- code: added null check to urlCommand
+
 ## [0.9.0] - 2021-04-16
 - Initial release lets you highlight text and right click for context menu to search